[Modify] 1. create app shops, users 2. divide models
