+++ .gitignore
@@ -362,4 +362,3 @@ zsdoc/data
 
 # End of https://www.toptal.com/developers/gitignore/api/python,pycharm,visualstudiocode,vim,macos,linux,zsh
 my_settings.py
-test
