+++ bagettt/settings.py
@@ -38,7 +38,9 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'corsheaders',
-    'packages'
+    'packages',
+    'users',
+    'shops',
 ]
 
 MIDDLEWARE = [
