+++ users/models.py
@@ -1,0 +1,13 @@
+from django.db import models
+
+class User(models.Model):
+    name         = models.CharField(max_length=50)
+    address      = models.CharField(max_length=200)
+    email        = models.CharField(max_length=200,unique=True)
+    password     = models.CharField(max_length=200)
+    phone_number = models.CharField(max_length=30)
+    created_at   = models.DateTimeField(auto_now_add=True)
+    updated_at   = models.DateTimeField(auto_now=True)
+
+    class Meta:
+        db_table = 'users'
